Add new dashboard

+++ src/UI/src/Expenses.tsx
@@ -1,12 +1,12 @@
 import React, { useState, useEffect, useMemo, useContext } from 'react'
 import { ColDef, GetRowIdFunc } from 'ag-grid-community'
-import { Expense } from './gensrc/Api'
 import { AgGridReact, CustomCellRendererProps } from 'ag-grid-react'
-import AutoCompleteCellEditor from './AutoCompleteCellEditor'
-import ButtonCellRenderer, { ButtonCellRendererProps } from './ButtonCellRenderer'
 import { Button, Spin } from 'antd'
 import { format } from 'date-fns'
+import AutoCompleteCellEditor from './AutoCompleteCellEditor'
+import ButtonCellRenderer, { ButtonCellRendererProps } from './ButtonCellRenderer'
 import { ApiContext, ExpenseCategoriesContext, PaymentMethodsContext, VendorsContext } from './Context'
+import { Expense } from './gensrc/Api'
 
 type ExpensesProps = {
   fromDate: Date
@@ -22,46 +22,37 @@ const Expenses = ({ fromDate, toDate }: ExpensesProps) => {
   const { expenseCategories } = useContext(ExpenseCategoriesContext)
   const { vendors, addVendor } = useContext(VendorsContext)
   const { paymentMethods, addPaymentMethod } = useContext(PaymentMethodsContext)
-  if (!expenseApi || !expenseCategories || !vendors || !paymentMethods) {
-    return (
-      <div style={{ display: 'flex', flex: 1, flexDirection: 'column' }}>
-        <Spin size="large" />
-      </div>
-    )
-  }
   const expenseCategoryIDToDisplayName: { [key: string]: string } = useMemo(
     () =>
-      expenseCategories.reduce(
+      expenseCategories?.reduce(
         (accumulator, { expenseCategoryID, displayName }) => ({
           ...accumulator,
           [expenseCategoryID!]: displayName,
         }),
         {} as { [key: string]: string }
-      ),
+      ) ?? {},
     [expenseCategories]
   )
-
   const paymentMethodIDToDisplayName: { [key: string]: string } = useMemo(
     () =>
-      paymentMethods.reduce(
+      paymentMethods?.reduce(
         (accumulator, { paymentMethodID, displayName }) => ({
           ...accumulator,
           [paymentMethodID!]: displayName,
         }),
         {} as { [key: string]: string }
-      ),
+      ) ?? {},
     [paymentMethods]
   )
-
   const vendorIDToDisplayName: { [key: string]: string } = useMemo(
     () =>
-      vendors.reduce(
+      vendors?.reduce(
         (accumulator, { vendorID, displayName }) => ({
           ...accumulator,
           [vendorID!]: displayName,
         }),
         {} as { [key: string]: string }
-      ),
+      ) ?? {},
     [vendors]
   )
 
@@ -69,7 +60,7 @@ const Expenses = ({ fromDate, toDate }: ExpensesProps) => {
   const [nextExpenseID, setNextExpenseID] = useState<number>(-1)
   useEffect(() => {
     expenseApi
-      .getExpenses({
+      ?.getExpenses({
         fromDate: format(fromDate, 'yyyy-MM-dd'),
         toDate: format(toDate, 'yyyy-MM-dd'),
       })
@@ -81,7 +72,7 @@ const Expenses = ({ fromDate, toDate }: ExpensesProps) => {
           }))
         )
       )
-  }, [fromDate, toDate])
+  }, [expenseApi, fromDate, toDate])
 
   const colDefs: ColDef<ExpenseRow>[] = useMemo(
     () => [
@@ -110,6 +101,9 @@ const Expenses = ({ fromDate, toDate }: ExpensesProps) => {
             )
           },
           clickedHandler: ({ data, api }: CustomCellRendererProps<ExpenseRow>) => {
+            if (!expenseApi) {
+              return
+            }
             const { original, date, amount, note, expenseCategoryID, vendorID, paymentMethodID, expenseID } = data!
             const payload: Expense = {
               expenseID: 0,
@@ -145,6 +139,9 @@ const Expenses = ({ fromDate, toDate }: ExpensesProps) => {
           text: 'Delete',
           isDisabledGetter: (expenseRow: ExpenseRow | undefined) => !expenseRow?.original,
           clickedHandler: ({ data, api }: CustomCellRendererProps<ExpenseRow>) => {
+            if (!expenseApi) {
+              return
+            }
             const promise = data!.original ? expenseApi.deleteExpense(data!.expenseID!) : Promise.resolve(void 0)
             promise.then(() => {
               api.applyTransaction({
@@ -219,6 +216,14 @@ const Expenses = ({ fromDate, toDate }: ExpensesProps) => {
       },
     ])
     setNextExpenseID(nextExpenseID - 1)
+  }
+
+  if (!expenseApi || !expenseCategories || !vendors || !paymentMethods) {
+    return (
+      <div style={{ display: 'flex', flex: 1, flexDirection: 'column' }}>
+        <Spin size="large" />
+      </div>
+    )
   }
 
   return (
